Changed idsite increments so they no longer increment by "1".
Right now Piwik just does a simple increment on the site ids.  If someone accidentally puts in a "5" rather than a "6" into ThoughtFarmer, they will add invalid data to someone else's statistics that would be very difficult to reverse.

It's not so much an obfuscation against malicious behaviour, as it is a defensive manoeuvre for a simple mistake.